ui(swush): ui makeover with swush branding (#11)
* feat: swush ui

* refactor: improve layout and spacing in Swap components

This commit refines the layout and spacing of the SwapContainer, SwapAction, and SwapHeader components to enhance the overall user interface. The adjustments include reducing padding and margin values, which streamlines the visual hierarchy and improves the alignment of elements. Specifically, the padding in the SwapContainer has been decreased from `pt-8` to `pt-4`, and the spacing in the SwapField has been adjusted to maintain consistency across the UI.

These changes not only contribute to a cleaner and more cohesive design but also ensure that the components remain responsive and visually appealing across different screen sizes. By focusing on modularity and maintainability, this refactor aligns with our commitment to producing high-quality, production-ready code while preserving existing functionality.

* ui(swush): enhance visual design of SwapConfirmSheet component

* ui(swush): enhance SwapField component with balance display and button size adjustment

* ui: add icons

* ui: override favicon

* ui: update layout spacing and icon disappear in responsive mode

* ui: add top margins for desktop and laptop screens

* ui: fix top margins for desktop and laptop screens

* ui: cleanup

* ui: revert top margins for desktop and laptop screens

* refactor: update input handling and balance display in Swap components

* fix: reset route when input is cleared or set to zero in SwapContainer

* fix: enhance input handling in SwapContainer and useSwapRoute

This commit improves the input handling logic in the SwapContainer component by canceling any pending debounced calls when the input is cleared or set to zero. Additionally, it updates the useSwapRoute hook to include safety checks that prevent API calls with invalid amounts, ensuring that only valid inputs trigger route fetching. These changes enhance the robustness of the swap functionality and prevent unnecessary API calls, ultimately improving user experience and performance.

* feat: add restart script

* ui: adjust layout spacing in SwapContainer for improved responsiveness